spi: cadence-qspi: Fix disable STIG mode
Fix commit# bc43217559:
Fix STIG mode reading from dev_get_driver_data by assigning to platdata
struct before read quirks variable.

+++ drivers/spi/cadence_qspi.h
@@ -270,7 +270,7 @@ struct cadence_spi_priv {
 
 struct cqspi_driver_platdata {
 	u32 hwcaps_mask;
-	u8 quirks;
+	u32 quirks;
 };
 
 /* Functions call declaration */
